physical_fitness.tsx moved to view directory
